optimize cross shard txs serialization
